extract `SignatureTypeArgumentProcessor` to separate file
so we can reuse it to import generic field types.
